Tool registry stage 4a: convert the 14 super-admin closure tools to a ToolDef domain
The last conversion batch of docs/TOOL-REGISTRY-DESIGN.md's strangler
migration — every remaining closure tool moves out of buildToolServer into
src/agent/tools/superAdmin.ts, bodies byte-verbatim, metadata mirrored from
rbac.ts/core.ts (cross-checked by tests/toolRegistry.test.ts), module state
moved exactly once:

- src/agent/tools/superAdmin.ts: grant_admin, revoke_admin, purge_user_data,
  audit_view, usage_stats, admin_activity, list_admins, engagement_stats,
  feature_flags, pause_bot, resume_bot, set_policy, redeploy_bot,
  suggest_issue (featureFlag (cfg) => cfg.github.enabled per core.ts's
  FEATURE_FLAGGED_TOOL_GROUPS; all minTier 'super_admin' per
  SUPER_ADMIN_TOOLS; none in DISCORD_ONLY_TOOLS so no platforms field) +
  ADMIN_DM_FAILED_NOTE, resetSessionsForRoleChange and reserveIssueDaily.
- buildToolServer now has ZERO closure tools: it is makeToolContext + the
  registry map on one flat createSdkMcpServer (117 tools, unchanged —
  pinned by toolRegistry.test.ts). The hand arrays in rbac.ts and core.ts's
  FEATURE_FLAGGED_TOOL_GROUPS still stand until the flip; the cross-check
  test keeps them honest.
- The two feature_flags SECURITY source scans in tests/tools.test.ts read
  tools.ts and anchored on the removed tool(...) wrapper shape — now scan
  superAdmin.ts and the defineTool handler shape, assertions unchanged,
  same repointing treatment stage 3 applied to the share_project scan.
